Move out rejoinAudio function to container
